Added Vector2D changed input of some functions to float. see comments for more details.

+++ SDLTemplate/Enemy.cpp
@@ -1,7 +1,7 @@
 #include "Enemy.h"
 #include "includes.h"
 
-void Enemy::load(int x, int y, int width, int height, std::string textureID)
+void Enemy::load(float x, float y, int width, int height, std::string textureID)
 {
 	GameObject::load(x, y, width, height, textureID);
 }
